cdk/requirements.txt
aws-cdk-lib==2.20.0
constructs>=10.0.0,<11.0.0
oe-patterns-cdk-common@git+https://github.com/ordinaryexperts/aws-marketplace-oe-patterns-cdk-common@c572af3ec4c8fe4b67b2241ff527ae753a6c84ed
